pyproject.toml
[tool.poetry]
name = "oauth2"
version = "0.1.0"
description = ""
authors = ["fsheee <fsheekhi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sqlmodel = "^0.0.19"
psycopg = {extras = ["binary"], version = "^3.1.19"}
httpx = "^0.27.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
